matches in tournaments dir
