+++ spec/support/openapi/response_helper.rb
@@ -1,7 +1,6 @@
 module OpenApi
   module Response
     JSON_CONTENT_TYPE = 'application/json'.freeze
-    NOT_FOUND = 'not found'.freeze
     def self.set_example_response_metadata
       lambda {
         after do |example|
